fix: reset video load state on back button click
This patches an issue where the video would start playing in the
background if a user clicked the back button before the video loaded.

Dividing the actual video player and load frame state may have to be
reconsidered. It may be easier to just include and maintain a custom
version of iced_video_player as part of this project. This only stops
the video from playing, but we are still loading the initial data in the background.